Changed mnxconverter to output JSON.
This updates the converter to output the newer JSON version
of MNX instead of XML.

There are still various bits and pieces to implement, such as:

* Clefs
* Tuplets
* Alternate ending durations
* Beams

...in order to get it to parity with the previous version of
the converter. But this commit does the bulk of the work in
setting up the infrastructure, and implementing specific notations
should be much less work going forward.